Allow overriding names of new windows and sessions.
This changeset modifies tmuxifier-load-window and tmuxifier-load-session
to accept an optional argument that sets the name of the new window or
session, overriding the default from the configuration file (or its
name).

This also required small changes to the `new_window()` and
`initialize_session()` functions in order to respect the `$window` and
`$session` variables, respectively, which were already set by the
calling `load_window()` and `load_session()` functions, but were
ignored.